test: add a semi-mt Alps test device
Provides the bounding box only, with slot 0 always being the upper/left, slot
1 being the lower-right touch. This needs to use the touch_down etc. litest
interfaces, which are now widened to double (leftover from 489630f58) and a
device-specific private pointer in the litest device.

New device feature for litest: LITEST_SEMI_MT

+++ test/litest-int.h
@@ -72,8 +72,8 @@ struct litest_test_device {
 };
 
 struct litest_device_interface {
-	void (*touch_down)(struct litest_device *d, unsigned int slot, int x, int y);
-	void (*touch_move)(struct litest_device *d, unsigned int slot, int x, int y);
+	void (*touch_down)(struct litest_device *d, unsigned int slot, double x, double y);
+	void (*touch_move)(struct litest_device *d, unsigned int slot, double x, double y);
 	void (*touch_up)(struct litest_device *d, unsigned int slot);
 
 	/**
